first check

+++ app/src/main/java/ru/myuniquenickname/myapplication/domain/entity/Genre.kt
@@ -1,3 +1,3 @@
 package ru.myuniquenickname.myapplication.domain.entity
 
-data class Genre(val id: Int, val name: String)
+data class Genre(val id: Long, val name: String)
